+++ src/App.js
@@ -1,12 +1,12 @@
 import "./App.css";
 import { LandingPage } from "./pages/LandingPage";
-import Login from "./pages/LoginPage";
-import PasswordReset from "./pages/PasswordResetPage";
+import LoginPage from "./pages/LoginPage";
+import PasswordResetPage from "./pages/PasswordResetPage";
 import SignupPage from "./pages/SignupPage";
 function App() {
   return (
     <div className="App">
-      <SignupPage />
+      <LandingPage />
     </div>
   );
 }
