+++ src/components/shared/Footer.jsx
@@ -4,7 +4,7 @@ import {
   faTwitter,
 } from "@fortawesome/free-brands-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import "../styles/footer.css";
+import "./footer.css";
 function Footer() {
   return (
     <>
